fonctions dans classes
