+++ navigation/localisation/include/localisation/LaserScan.h
