+++ navigation/localisation/src/EKF_node_with_lines.cpp
@@ -9,7 +9,7 @@
 #include "geometry_msgs/Pose2D.h"
 #include "geometry_msgs/Twist.h"
 #include "nav_msgs/Odometry.h"
-#include "laserScan.h"
+#include "LaserScan.h"
 #include "landmarks_detection_utils.h"
 
 #include "EKF_class.h"
