+++ navigation/localisation/src/try.cpp
@@ -11,7 +11,7 @@
 #include "geometry_msgs/Pose2D.h"
 #include "geometry_msgs/Twist.h"
 #include "nav_msgs/Odometry.h"
-#include "laserScan.h"
+#include "LaserScan.h"
 #include "landmarks_detection_utils.h"
 #include "math_functions.h"
 #include "Segment.h"
